- Clean a bunch of compilation warning related to 'possible loss of data' between numeric types. - Entities now hold a reference to the scene they belong to. - Add a state flags var to Entity. - Refactor 'active' to 'running' Scene. - Canvas adds its camera to the Scene it belongs to instead of the current running Scene. - Entity::OnStart(...) is called properly when an Entity enter in a running scene graph and is not started yet.

+++ vongine/src/vongine/ui/VGUIMessages.h
@@ -21,8 +21,8 @@ namespace ui
 
 	struct DLLAPI MessageMouseButtonDown : public Message
 	{
-		MessageMouseButtonDown() : MessageMouseButtonDown(0, 0.f, 0.f) {}
-		MessageMouseButtonDown(const uint8 _buttonId, const float _x, const float _y)
+		MessageMouseButtonDown() : MessageMouseButtonDown(0, 0, 0) {}
+		MessageMouseButtonDown(const uint8 _buttonId, const uint32 _x, const uint32 _y)
 		{
 			type = MessageType::MouseButtonDown;
 			x = _x;
@@ -30,14 +30,14 @@ namespace ui
 		}
 
 		uint8 buttonId;
-		float x;
-		float y;
+		uint32 x;
+		uint32 y;
 	};
 
 	struct DLLAPI MessageMouseButtonUp : public Message
 	{
-		MessageMouseButtonUp() : MessageMouseButtonUp(0, 0.f, 0.f) {}
-		MessageMouseButtonUp(const uint8 _buttonId, const float _x, const float _y)
+		MessageMouseButtonUp() : MessageMouseButtonUp(0, 0, 0) {}
+		MessageMouseButtonUp(const uint8 _buttonId, const uint32 _x, const uint32 _y)
 		{
 			type = MessageType::MouseButtonUp;
 			x = _x;
@@ -45,8 +45,8 @@ namespace ui
 		}
 
 		uint8 buttonId;
-		float x;
-		float y;
+		uint32 x;
+		uint32 y;
 	};
 }
